Overhaul reading to use go readers
Reading savedata is now a 1-stage process since nobody needs the
header data.
Privdump benefits greatly from losing its own form-loading code.
rip.go also loses its own form-loading code

Add some parser functions to parse bytes that we have already read.

+++ tables/tables.go
@@ -142,7 +142,7 @@ func Systems(gt types.Game) map[SYS_ID]Sysinfo {
 	return systems
 }
 
-var Cargo = map[uint8]string{
+var Cargo = map[int]string{
 	0:  "Grain",
 	1:  "Generic Foods",
 	2:  "Luxury Foods",
